+++ services/web/server/tests/unit/isolated/test_studies_dispatcher_core.py
@@ -1,20 +1,23 @@
-# pylint:disable=unused-variable
-# pylint:disable=unused-argument
-# pylint:disable=redefined-outer-name
-# from simcore_postgres_database.models.projects import projects as projects_table
+# pylint: disable=redefined-outer-name
+# pylint: disable=unused-argument
+# pylint: disable=unused-variable
+# pylint: disable=too-many-arguments
+
 
 import json
 import urllib.parse
 from pathlib import Path
-from typing import Dict
 from urllib.parse import parse_qs
 
 import pytest
-from models_library.projects import Project
+from aiohttp.test_utils import make_mocked_request
+from models_library.projects import Project, ProjectID
+from models_library.projects_nodes_io import NodeID
 from pydantic import ConfigError, ValidationError, validator
 from pydantic.main import BaseModel
 from pydantic.networks import HttpUrl
 from pytest_simcore.helpers.utils_services import list_fake_file_consumers
+from servicelib.aiohttp.requests_validation import parse_request_query_parameters_as
 from simcore_service_webserver._constants import APP_JSONSCHEMA_SPECS_KEY
 from simcore_service_webserver.projects import projects_api
 from simcore_service_webserver.studies_dispatcher._projects import (
@@ -29,21 +32,25 @@ from yarl import URL
 
 
 @pytest.fixture
-def project_jsonschema(project_schema_file: Path) -> Dict:
+def project_jsonschema(project_schema_file: Path) -> dict:
     return json.loads(project_schema_file.read_text())
 
 
 def test_download_link_validators():
 
-    params = RedirectionQueryParams(
-        file_name="dataset_description.slsx",
-        file_size=24770,  # BYTES
-        file_type="MSExcel",
-        viewer_key="simcore/services/dynamic/fooo",
-        viewer_version="1.0.0",
-        download_link="https://blackfynn-discover-use1.s3.amazonaws.com/23/2/files/dataset_description.xlsx?AWSAccessKeyId=AKIAQNJEWKCFAOLGQTY6&Signature=K229A0CE5Z5OU2PRi2cfrfgLLEw%3D&x-amz-request-payer=requester&Expires=1605545606",
+    params = RedirectionQueryParams.parse_obj(
+        dict(
+            file_name="dataset_description.slsx",
+            file_size=24770,  # BYTES
+            file_type="MSExcel",
+            viewer_key="simcore/services/dynamic/fooo",
+            viewer_version="1.0.0",
+            download_link="https://blackfynn-discover-use1.s3.amazonaws.com/23/2/files/dataset_description.xlsx?AWSAccessKeyId=AKIAQNJEWKCFAOLGQTY6&Signature=K229A0CE5Z5OU2PRi2cfrfgLLEw%3D&x-amz-request-payer=requester&Expires=1605545606",
+        )
     )
-    assert params.download_link.host.endswith("s3.amazonaws.com")
+    assert params.download_link.host and params.download_link.host.endswith(
+        "s3.amazonaws.com"
+    )
     assert params.download_link.host_type == "domain"
 
     # TODO: cannot validate amazon download links with HEAD because only operation allowed is GET
@@ -65,9 +72,9 @@ async def test_create_project_with_viewer(project_jsonschema, view):
     user = UserInfo(id=3, name="foo", primary_gid=33, email="foo@bar.com")
 
     project = create_viewer_project_model(
-        project_id="e3ee7dfc-25c3-11eb-9fae-02420a01b846",
-        file_picker_id="4c69c0ce-00e4-4bd5-9cf0-59b67b3a9343",
-        viewer_id="fc718e5a-bf07-4abe-b526-d9cafd34830c",
+        project_id=ProjectID("e3ee7dfc-25c3-11eb-9fae-02420a01b846"),
+        file_picker_id=NodeID("4c69c0ce-00e4-4bd5-9cf0-59b67b3a9343"),
+        viewer_id=NodeID("fc718e5a-bf07-4abe-b526-d9cafd34830c"),
         owner=user,
         download_link="http://httpbin.org/image/jpeg",
         viewer_info=viewer,
@@ -78,12 +85,14 @@ async def test_create_project_with_viewer(project_jsonschema, view):
     assert list(project.workbench.keys())
 
     # converts into equivalent Dict
-    project_in: Dict = json.loads(project.json(exclude_none=True, by_alias=True))
+    project_in: dict = json.loads(project.json(exclude_none=True, by_alias=True))
     print(json.dumps(project_in, indent=2))
 
     # This operation is done exactly before adding to the database in projects_handlers.create_projects
     await projects_api.validate_project(
-        app={APP_JSONSCHEMA_SPECS_KEY: {"projects": project_jsonschema}},
+        app={  # type: ignore
+            APP_JSONSCHEMA_SPECS_KEY: {"projects": project_jsonschema},
+        },
         project=project_in,
     )
 
@@ -101,15 +110,13 @@ async def test_create_project_with_viewer(project_jsonschema, view):
         ),
     ],
 )
-def test_validation_of_redirection_urls(url_in, expected_download_link):
+def test_validation_of_redirection_urls(url_in: str, expected_download_link: str):
     #
     #  %  -> %25
     #  space -> %20
     #
-    url_in = URL(url_in)
-
-    assert hasattr(url_in, "query")  # as a Request
-    params = RedirectionQueryParams.from_request(url_in)
+    mock_request = make_mocked_request(method="GET", path=f"{URL(url_in).relative()}")
+    params = parse_request_query_parameters_as(RedirectionQueryParams, mock_request)
 
     assert params.download_link == expected_download_link
