+++ services/web/server/tests/unit/with_dbs/01/studies_dispatcher/test_studies_dispatcher_handlers.py
@@ -2,7 +2,6 @@
 # pylint:disable=unused-argument
 # pylint:disable=redefined-outer-name
 
-import logging
 import re
 import urllib.parse
 from pprint import pprint
@@ -16,71 +15,18 @@ from aioresponses import aioresponses
 from models_library.projects_state import ProjectLocked, ProjectStatus
 from pydantic import parse_obj_as
 from pytest import MonkeyPatch
-from pytest_simcore.helpers.typing_env import EnvVarsDict
 from pytest_simcore.helpers.utils_assert import assert_status
-from pytest_simcore.helpers.utils_envs import setenvs_from_dict
 from pytest_simcore.helpers.utils_login import UserRole
 from settings_library.redis import RedisSettings
 from simcore_service_webserver import catalog
-from simcore_service_webserver.log import setup_logging
 from simcore_service_webserver.studies_dispatcher._core import ViewerInfo
 from simcore_service_webserver.studies_dispatcher.handlers_rest import ServiceGet
-from simcore_service_webserver.studies_dispatcher.settings import (
-    StudiesDispatcherSettings,
-)
 from sqlalchemy.sql import text
 from yarl import URL
 
 #
 # FIXTURES OVERRIDES -----------------------------------------------------------------------------------------------
 #
-
-
-@pytest.fixture
-def app_environment(app_environment: EnvVarsDict, monkeypatch: MonkeyPatch):
-    envs_plugins = setenvs_from_dict(
-        monkeypatch,
-        {
-            "WEBSERVER_ACTIVITY": "null",
-            "WEBSERVER_CATALOG": "null",
-            "WEBSERVER_CLUSTERS": "null",
-            "WEBSERVER_COMPUTATION": "0",
-            "WEBSERVER_DIAGNOSTICS": "null",
-            "WEBSERVER_DIRECTOR": "null",
-            "WEBSERVER_EXPORTER": "null",
-            "WEBSERVER_GROUPS": "1",
-            "WEBSERVER_META_MODELING": "null",
-            "WEBSERVER_PRODUCTS": "1",
-            "WEBSERVER_PUBLICATIONS": "0",
-            "WEBSERVER_RABBITMQ": "null",
-            "WEBSERVER_REMOTE_DEBUG": "0",
-            "WEBSERVER_SOCKETIO": "0",
-            "WEBSERVER_STORAGE": "null",
-            "WEBSERVER_TAGS": "1",
-            "WEBSERVER_TRACING": "null",
-            "WEBSERVER_USERS": "1",
-            "WEBSERVER_VERSION_CONTROL": "0",
-        },
-    )
-
-    monkeypatch.delenv("WEBSERVER_STUDIES_DISPATCHER", raising=False)
-    app_environment.pop("WEBSERVER_STUDIES_DISPATCHER", None)
-
-    envs_studies_dispatcher = setenvs_from_dict(
-        monkeypatch,
-        {
-            "STUDIES_ACCESS_ANONYMOUS_ALLOWED": "1",
-            "STUDIES_GUEST_ACCOUNT_LIFETIME": "2 1:10:00",  # 2 days 1h and 10 mins
-        },
-    )
-
-    # NOTE: To see logs, use pytest -s --log-cli-level=DEBUG
-    setup_logging(level=logging.DEBUG)
-
-    plugin_settings = StudiesDispatcherSettings.create_from_envs()
-    print(plugin_settings.json(indent=1))
-
-    return {**app_environment, **envs_plugins, **envs_studies_dispatcher}
 
 
 @pytest.fixture
